+++ src/modules/user/user.service.ts
@@ -1,5 +1,5 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { CreateUserDto, UserLoginDto } from './user.dto';
+import { CreateUserDto, UserDto, UserLoginDto } from './user.dto';
 import * as bcrypt from 'bcrypt';
 import { UserDocument } from './user.interface';
 import { InjectModel } from '@nestjs/mongoose';
@@ -10,6 +10,7 @@ import { BadRequestException } from '../../exceptions/bad-request.exception';
 import { LoggerService } from '../logger/logger.service';
 import { NotFoundException } from '../../exceptions/not-found.exception';
 import { Redis } from 'ioredis';
+import { CapabilityType } from './user.types';
 
 @Injectable()
 export class UserService {
@@ -22,13 +23,7 @@ export class UserService {
     ) {}
 
     async getUser(login: string): Promise<UserDocument> {
-        const user = await this.userModel.findOne({ login });
-
-        if (!user) {
-            throw new NotFoundException('UserService/getUser', 'User does not found');
-        }
-
-        return user;
+        return this.userModel.findOne({ login });
     }
 
     async loginUser(userLoginDto: UserLoginDto, res): Promise<UserDocument> {
@@ -65,7 +60,7 @@ export class UserService {
     }
 
     async createUser(createUserDto: CreateUserDto): Promise<UserDocument> {
-        const context = 'UserService/createUser:';
+        const context = 'UserService/createUser';
 
         if (await this.getUser(createUserDto.login)) {
             const message = `User with "${createUserDto.login}" login has already exists`;
@@ -92,5 +87,75 @@ export class UserService {
 
     async areSameHashedPasswords(password: string, hashedPassword: string): Promise<boolean> {
         return await bcrypt.compare(password, hashedPassword);
+    }
+
+    async grantPermission(user: UserDto, byUser: UserDto, capability: CapabilityType): Promise<boolean> {
+        const context = 'UserService/grantPermission';
+
+        if (!user) {
+            const message = 'Failed action to grant a permission. User with provided login does not exist.';
+            this.loggerService.error(context, message);
+
+            throw new BadRequestException(context, message);
+        }
+
+        if (user.capabilities && user.capabilities[capability]) {
+            this.loggerService.info(context, `User "${user.login}" has provided capability.`);
+
+            return false;
+        }
+
+        await this.userModel.updateOne(
+            {
+                _id: user._id,
+                login: user.login
+            },
+            {
+                $set: {
+                    capabilities: {
+                        ...user.capabilities,
+                        [capability]: true
+                    }
+                }
+            }
+        );
+        this.loggerService.info(context, `User "${byUser.login}" has granted permission "${capability}" to "${user.login}" user.`);
+
+        return true;
+    }
+
+    async denyPermission(user: UserDto, byUser: UserDto, capability: CapabilityType): Promise<boolean> {
+        const context = 'UserService/denyPermission';
+
+        if (!user) {
+            const message = 'Failed action to deny a permission. User with provided login does not exist.';
+            this.loggerService.error(context, message);
+
+            throw new BadRequestException(context, message);
+        }
+
+        if (!user.capabilities || !user.capabilities[capability]) {
+            this.loggerService.info(context, `User "${user.login}" has not provided capability.`);
+
+            return false;
+        }
+
+        const newCapabilities = user.capabilities;
+        delete newCapabilities[capability];
+
+        await this.userModel.updateOne(
+            {
+                _id: user._id,
+                login: user.login
+            },
+            {
+                $set: {
+                    capabilities: newCapabilities
+                }
+            }
+        );
+        this.loggerService.info(context, `User "${byUser.login}" has denied permission "${capability}" to "${user.login}" user.`);
+
+        return true;
     }
 }